Extract timeline persistence and item placement helpers
